Add files via upload
The Pomodoro timer is a productivity tool that helps individuals manage their time effectively by breaking work into intervals, traditionally 25 minutes in length, separated by short breaks. This technique is based on the idea that frequent breaks can improve mental agility and focus.

The timer operates in a simple yet effective manner: it counts down the time for each work interval, alerting the user when the time is up. After completing a work interval, it prompts the user to take a short break. This cycle repeats until a longer break is taken after several work intervals.

The Pomodoro timer program presented here is a digital implementation of this technique. It features a user-friendly interface with a large timer display, clear start  buttons, and a customizable background color. Users can easily start  the timer, adjust the work interval duration, and track their progress throughout the day.

Additionally, the program includes a feature to write any journals and add our to-do lists. This simple yet elegant design aims to enhance the user's focus and productivity while providing a visually appealing and intuitive interface.